tests/wpt/tests/css/css-transitions/shadow-root-insertion.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transitions: behavior when a shadow root is inserted while transitioning</title>
<meta name="assert" content="Checks the addition of a shadow root does not affect an in-flight transition">
<link rel="help" href="https://drafts.csswg.org/css-transitions/">

<script src="/resources/testharness.js" type="text/javascript"></script>
<script src="/resources/testharnessreport.js" type="text/javascript"></script>
<script src="./support/helper.js" type="text/javascript"></script>

</head>
<body>
<div id="log"></div>
<script>
test(t => {
  // Start a 100s transition 50% of the way through
  const div = addDiv(t, {
    style: 'transition: height 100s -50s linear; height: 0px',
  });
  getComputedStyle(div).height;
  div.style.height = '100px';
  assert_equals(
    getComputedStyle(div).height,
    '50px',
    'Transition should be initially 50% complete'
  );

  // Add a shadow root
  div.attachShadow({ mode: "open" });

  // The transition on the height property should not have been canceled
  assert_equals(
    getComputedStyle(div).height,
    '50px',
    'Transition should not have been canceled'
  );
}, 'addition of a shadow root should not cancel in-flight transitions');
</script>

</body>
</html>
